Start cleaning up implementation
