+++ modules/core/src/main/scala/test.scala
@@ -5,7 +5,10 @@ import monocle._
 
 import shapeless.syntax.singleton._
 
-object module extends JsonModule[JsonSchema.type] with HasMigration[JsonSchema.type] {
+object module
+    extends JsonModule[JsonSchema.type]
+    with HasMigration[JsonSchema.type]
+    with Migrations[JsonSchema.type] {
 
 //  import Transform._
 
@@ -35,133 +38,8 @@ object module extends JsonModule[JsonSchema.type] with HasMigration[JsonSchema.t
     )
   val homer = Person("Homer Simpson", Some(User(active = true, username = "SpiderPig")))
 
-  trait Lookup[Re, A] {
-    def apply(registry: Re): Schema[A]
-  }
-
-  trait LowPrioLookup {
-    implicit def rightLookup[R1, RT, A](implicit rest: Lookup[RT, A]): Lookup[(R1, RT), A] =
-      new Lookup[(R1, RT), A] {
-        def apply(registry: (R1, RT)): Schema[A] = rest(registry._2)
-      }
-  }
-
-  object Lookup extends LowPrioLookup {
-    implicit def leftLookup[RR, A]: Lookup[(Schema[A], RR), A] =
-      new Lookup[(Schema[A], RR), A] {
-        def apply(registry: (Schema[A], RR)): Schema[A] = registry._1
-      }
-  }
-
-  trait Replace[Tpe, Re, A] {
-    def apply(registry: Re, replacement: Ctr[Tpe, A]): Re
-  }
-
-  object Replace {
-    implicit def rightReplace[Tpe, A, RT]: Replace[Tpe, (Ctr[Tpe, A], RT), A] =
-      new Replace[Tpe, (Ctr[Tpe, A], RT), A] {
-        override def apply(
-          registry: (Ctr[Tpe, A], RT),
-          replacement: Ctr[Tpe, A]
-        ): (Ctr[Tpe, A], RT) = (replacement, registry._2)
-      }
-
-    implicit def leftReplace[Tpe, A, R1, RT](
-      implicit rest: Replace[Tpe, RT, A]
-    ): Replace[Tpe, (R1, RT), A] =
-      new Replace[Tpe, (R1, RT), A] {
-        override def apply(
-          registry: (R1, RT),
-          replacement: Ctr[Tpe, A]
-        ): (R1, RT) =
-          (registry._1, rest(registry._2, replacement))
-      }
-  }
-
-  type Ctr[Re, A] = Registry[Re] => Schema[A]
-
-  case class CtrRegistry[Types, Re](registry: Re) {
-
-    def addEntry[A](ctr: Ctr[Types, A]): CtrRegistry[(Schema[A], Types), (Ctr[Types, A], Re)] =
-      new CtrRegistry[(Schema[A], Types), (Ctr[Types, A], Re)]((ctr, registry))
-
-    def replace[Tpe, A](
-      replacement: Ctr[Tpe, A]
-    )(implicit replace: Replace[Tpe, Re, A]): CtrRegistry[Types, Re] =
-      new CtrRegistry(replace(registry, replacement))
-  }
-
-  trait Build[Tpe, Reg] {
-    def apply(ctrReg: CtrRegistry[Tpe, Reg]): Registry[Tpe]
-  }
-
-  object Build {
-    implicit val buildEmpty: Build[Unit, Unit] = new Build[Unit, Unit] {
-      def apply(ctrReg: CtrRegistry[Unit, Unit]): Registry[Unit] = new Registry(ctrReg.registry)
-    }
-
-    implicit def buildRegistry[A, Tpe, Reg](
-      implicit rest: Build[Tpe, Reg]
-    ): Build[(Schema[A], Tpe), (Ctr[Tpe, A], Reg)] =
-      new Build[(Schema[A], Tpe), (Ctr[Tpe, A], Reg)] {
-
-        def apply(
-          ctrReg: CtrRegistry[(Schema[A], Tpe), (Ctr[Tpe, A], Reg)]
-        ): Registry[(Schema[A], Tpe)] = {
-          val tail = rest(new CtrRegistry[Tpe, Reg](ctrReg.registry._2))
-
-          new Registry((ctrReg.registry._1(tail), tail.registry))
-        }
-      }
-  }
-
-  object CtrRegistry {
-    def empty: CtrRegistry[Unit, Unit] = new CtrRegistry(())
-
-    def build[Tpe, Reg](ctrReg: CtrRegistry[Tpe, Reg])(implicit b: Build[Tpe, Reg]): Registry[Tpe] =
-      b(ctrReg)
-  }
-
-  case class Registry[Re](registry: Re) { self =>
-
-    def lookup[A](implicit l: Lookup[Re, A]): Schema[A] = l(registry)
-
-    def addEntry[A](mkEntry: Registry[Re] => Schema[A]): Registry[(Schema[A], Re)] =
-      new Registry((mkEntry(self), registry))
-
-  }
-
-  object Registry {
-    val empty: Registry[Unit] = new Registry(())
-
-  }
-
-  trait Provides[A] {
-    type T[Re] = Lookup[Re, A]
-  }
-
-  /*
-  trait Changelog[X, RA, A] {}
-  final case class Init[X, RA, A](create: X => SchemaZ[RA, A]) extends Changelog[X, RA, A]
-  final case class Migrate[X, RA, RX, A](
-    base: Changelog[X, RA, A],
-    mig: SchemaZ[RA, A] => SchemaZ[RX, A]
-  ) extends Changelog[X, RX, A]
-
-  object Migrate {
-
-    def create[X, RA, RX, A](base: Changelog[X, RA, A])(mig: SchemaZ[RA, A] => SchemaZ[RX, A]) =
-      Migrate(base, mig)
-  }
-
-  final case class Prodm[X0, X1, R0, R1, A0, A1, RX, X](
-    l: Changelog[X0, R0, A0],
-    r: Changelog[X1, R1, A1],
-    combine: Changelog[(SchemaZ[R0, A0], SchemaZ[R1, A1]), RX, X]
-  ) extends Changelog[(X0, X1), RX, X]
-   */
-  def mkUser =
-    (_: Registry[Unit]) =>
+  def mkUser[Re] =
+    (_: Registry[Re]) =>
       record(
         "active".narrow -*>: prim(JsonSchema.JsonBool) :*: "username".narrow -*>: prim(
           JsonSchema.JsonString
@@ -169,18 +47,16 @@ object module extends JsonModule[JsonSchema.type] with HasMigration[JsonSchema.t
         Iso[(Boolean, String), User]((User.apply _).tupled)(u => (u.active, u.username))
       )
 
-//  val initU = Init(mkUser)
-//  val v1    = Migrate.create(initU)(_.addField("active".narrow, false))
+  def mkAdmin[Re] =
+    (_: Registry[Re]) =>
+      record(
+        "username".narrow -*>: prim(JsonSchema.JsonString) :*: "rights".narrow -*>: seq(
+          prim(JsonSchema.JsonString)
+        ),
+        Iso[(String, List[String]), Admin]((Admin.apply _).tupled)(a => (a.username, a.rights))
+      )
 
-  val mkAdmin =
-    record(
-      "username".narrow -*>: prim(JsonSchema.JsonString) :*: "rights".narrow -*>: seq(
-        prim(JsonSchema.JsonString)
-      ),
-      Iso[(String, List[String]), Admin]((Admin.apply _).tupled)(a => (a.username, a.rights))
-    )
-
-  def mkRole[Re: Provides[User]#T: Provides[Admin]#T] =
+  def mkRole[Re: Lookup[?, User]: Lookup[?, Admin]] =
     (r: Registry[Re]) =>
       union(
         "user".narrow -+>: r.lookup[User] :+:
@@ -194,9 +70,7 @@ object module extends JsonModule[JsonSchema.type] with HasMigration[JsonSchema.t
         }
       )
 
-  type WithRegistry[A, RA, Re] = Registry[Re] => SchemaZ[RA, A]
-
-  def mkPerson[Re: Provides[Role]#T] =
+  def mkPerson[Re: Lookup[?, Role]] =
     (reg: Registry[Re]) =>
       record(
         "name".narrow -*>: prim(JsonSchema.JsonString) :*:
@@ -208,17 +82,29 @@ object module extends JsonModule[JsonSchema.type] with HasMigration[JsonSchema.t
 
   val version2 = CtrRegistry.empty
     .addEntry(mkUser)
-    .addEntry(_ => mkAdmin)
+    .addEntry(mkAdmin)
     .addEntry(mkRole)
     .addEntry(mkPerson)
 
-  val version2Reg = CtrRegistry.build(version2)
+  val version2Reg = Registry.build(version2)
 
-  def userV1 = mkUser.andThen(_.addField("active".narrow, false))
+  def userV1[Re] =
+    (r: Registry[Re]) => mkUser(r).addField("active".narrow, false)
+
+  def adminV1[Re] = (r: Registry[Re]) => mkAdmin(r).addField("username".narrow, "John Doe")
 
   val version1 = version2
-    .replace(userV1)
-//    .addEntry(mkRole)
+    .migrate[User]
+    .apply(userV1)
+    .migrate[Admin]
+    .apply(adminV1)
 
-  // val version1Reg = CtrRegistry.build(version1)
+  val version1Reg = Registry.build(version1)
+
+  def userV0[Re] = (r: Registry[Re]) => userV1(r).addField("username".narrow, "John Doe")
+
+  val version0 = version1.migrate[User].apply(userV0)
+
+  val version0Reg = Registry.build(version0)
+
 }
